requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
bidict==0.23.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.135.1
fastapi-socketio==0.0.10
h11==0.16.0
idna==3.11
iniconfig==2.3.0
motor==3.7.1
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-jose==3.5.0
python-socketio==5.16.1
redis==7.3.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wsproto==1.3.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0